Decouple file upload from project details
