+++ src/utils/upload/processUpload.js
@@ -14,7 +14,7 @@ const putInS3 = async (bucketKey, loadedFileData, dispatch, sampleUuid, fileName
       metadata,
       progressCallback(progress) {
         const percentProgress = Math.round((progress.loaded / progress.total) * 100);
-
+        console.log('Calling from putInS3');
         dispatch(updateSampleFile(sampleUuid, fileName, {
           upload: {
             status: UploadStatus.UPLOADING,
@@ -45,18 +45,19 @@ const compressAndUploadSingleFile = async (
   let loadedFile = null;
 
   try {
-    loadedFile = await loadAndCompressIfNecessary(file, () => (
-      dispatch(
+    loadedFile = await loadAndCompressIfNecessary(file, () => {
+      console.log('Calling from compressAndUploadSingleFile try 1');
+      return dispatch(
         updateSampleFile(
           sampleUuid,
           fileName,
           { upload: { status: UploadStatus.COMPRESSING } },
         ),
-      )
-    ));
+      );
+    });
   } catch (e) {
     const fileErrorStatus = e === 'aborted' ? UploadStatus.FILE_READ_ABORTED : UploadStatus.FILE_READ_ERROR;
-
+    console.log('Calling from compressAndUploadSingleFile error');
     dispatch(
       updateSampleFile(
         sampleUuid,
@@ -73,7 +74,7 @@ const compressAndUploadSingleFile = async (
       bucketKey, loadedFile, dispatch,
       sampleUuid, fileName, metadata,
     );
-
+    console.log('Calling from compressAndUploadSingleFile try 2');
     dispatch(
       updateSampleFile(
         sampleUuid,
@@ -87,6 +88,7 @@ const compressAndUploadSingleFile = async (
 
     await uploadPromise;
   } catch (e) {
+    console.log('Calling from compressAndUploadSingleFile error 2');
     dispatch(
       updateSampleFile(
         sampleUuid,
@@ -98,6 +100,7 @@ const compressAndUploadSingleFile = async (
     return;
   }
 
+  console.log('Calling again because why not');
   dispatch(
     updateSampleFile(
       sampleUuid,
@@ -142,6 +145,7 @@ const compressAndUpload = (sample, activeProjectUuid, dispatch) => Object.fromEn
 );
 
 const processUpload = async (filesList, sampleType, samples, activeProjectUuid, dispatch) => {
+  console.log('MADE IT !!!!!!');
   const samplesMap = filesList.reduce((acc, file) => {
     const pathToArray = file.name.trim().replace(/[\s]{2,}/ig, ' ').split('/');
 
@@ -177,6 +181,7 @@ const processUpload = async (filesList, sampleType, samples, activeProjectUuid, 
     sample.files = compressAndUpload(sample, activeProjectUuid, dispatch);
 
     Object.values(sample.files).forEach((file) => {
+      console.log('Calling from processUpload');
       // Create files
       dispatch(updateSampleFile(sample.uuid, file.name, {
         ...file,
